ServerlessComputing/GoogleCloud/src/requirements.txt
pillow
flask==1.1.1